+++ changelog.md
@@ -10,3 +10,4 @@
 - Add persist task state
 - Add multinode support
 - Add container monitor to restart if a container failed
+- Add RestCall to reflate missing zookeeper and kafka processes, for the case the monitoring can not find the problem